# OOCL.Client/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "OOCL.Api",
		"version": "1.0"
	},
	"paths": {
...
		"/api/Audio/audios/{guid}/info": {
			"get": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/Image/images/{guid}/image64/{format}": {
			"get": {
				"tags": [
					"Image"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "format",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "bmp"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ImageData"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageData"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageData"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/OpenCl/status": {
			"get": {
				"tags": [
					"OpenCl"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/OpenCl/devices": {
			"get": {
				"tags": [
					"OpenCl"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClDeviceInfo"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClDeviceInfo"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClDeviceInfo"
									}
								}
							}
						}
					},
					"204": {
						"description": "No Content",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClDeviceInfo"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClDeviceInfo"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClDeviceInfo"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/OpenCl/devices/{deviceId}/initialize": {
			"post": {
				"tags": [
					"OpenCl"
				],
				"parameters": [
					{
						"name": "deviceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 2
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/OpenCl/dispose": {
			"delete": {
				"tags": [
					"OpenCl"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClServiceInfo"
								}
							}
						}
					}
				}
			}
		},
		"/api/OpenCl/usage": {
			"get": {
				"tags": [
					"OpenCl"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/OpenClUsageInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClUsageInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClUsageInfo"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/OpenClUsageInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClUsageInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/OpenClUsageInfo"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/OpenCl/memory": {
			"get": {
				"tags": [
					"OpenCl"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							}
						}
					},
					"204": {
						"description": "No Content",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OpenClMemoryInfo"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/OpenCl/executeMandelbrot/{kernel}/{version}/{width}/{height}/{zoom}/{x}/{y}/{coeff}/{r}/{g}/{b}/{copyGuid}/{allowTempSession}": {
			"get": {
				"tags": [
					"OpenCl"
				],
				"parameters": [
					{
						"name": "kernel",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "mandelbrotPrecise"
						}
					},
					{
						"name": "version",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"default": "01"
						}
					},
					{
						"name": "width",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 1920
						}
					},
					{
						"name": "height",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 1080
						}
					},
					{
						"name": "zoom",
						"in": "path",
						"required": true,
						"schema": {
							"type": "number",
							"format": "double",
							"default": 1
						}
					},
					{
						"name": "x",
						"in": "path",
						"required": true,
						"schema": {
							"type": "number",
							"format": "double",
							"default": 0
						}
					},
					{
						"name": "y",
						"in": "path",
						"required": true,
						"schema": {
							"type": "number",
							"format": "double",
							"default": 0
						}
					},
					{
						"name": "coeff",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 8
						}
					},
					{
						"name": "r",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 0
						}
					},
					{
						"name": "g",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 0
						}
					},
					{
						"name": "b",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 0
						}
					},
					{
						"name": "copyGuid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "boolean",
							"default": true
						}
					},
					{
						"name": "allowTempSession",
						"in": "path",
						"required": true,
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ImageObjInfo"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AudioObjInfo": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid",
						"readOnly": true
					},
					"name": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"filepath": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"meta": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"onHost": {
						"type": "boolean",
						"readOnly": true
					},
					"sizeMb": {
						"type": "number",
						"format": "float",
						"readOnly": true
					},
					"pointerHex": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataType": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataStructure": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"samplerate": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"channels": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"bitdepth": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"lenght": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"chunkSize": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"overlapSize": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"form": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"stretchFactor": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"playing": {
						"type": "boolean",
						"readOnly": true
					},
					"playbackState": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"volume": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"duration": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"currentTime": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"lastLoadingTime": {
						"type": "number",
						"format": "double"
					},
					"lastProcessingTime": {
						"type": "number",
						"format": "double"
					},
					"errorMessage": {
						"type": "string",
						"nullable": true
					},
					"bpm": {
						"type": "number",
						"format": "float"
					}
				},
				"additionalProperties": false
			},
...
			"ProblemDetails": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "integer",
						"format": "int32",
						"nullable": true
					},
					"detail": {
						"type": "string",
						"nullable": true
					},
					"instance": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": {}
			}
		}
	}
}